+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.16.0
+* Deprecate `GrantedPaymentInstrumentUpdate` and add `GrantorUpdatedGrantedPaymentMethod` and `RecipientUpdatedGrantedPaymentMethod`
+* Add `accountType` to `Transaction`, `PaymentMethod`, and `CreditCardVerification`.
+
 ## 2.15.0
 * Add `bin` field to `paymentMethodNonceDetails`
 * Add Error indicating pdf uploads too long for dispute evidence.